preseed: add software RAID definition support
The debian-installer preseed template only supported to create LVM
physical volume on one block device. With this commit, it is now
possible to define software RAID on multiple block devices. The
previous behaviour becomes the default behaviour, existing
installations just have to define part_method: lvm in the defaults
section of the boot-params.yaml.

To enable software RAID feature, the part_method must be set to `raid`
and the part_raid must be defined with the partman-auto-raid recipe.